+++ iOS/Projects/Features/Home/Sources/Presntaion/HomeScene/ViewController/HomeViewController.swift
@@ -175,20 +175,6 @@ private extension HomeViewController {
   enum Metrics {}
 }
 
-private extension HomeViewController {
-  static func makeFeedCollectionViewLayout() -> UICollectionViewCompositionalLayout {
-    let itemSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(1), heightDimension: .fractionalHeight(1))
-    let item = NSCollectionLayoutItem(layoutSize: itemSize)
-    item.contentInsets = .init(top: 9, leading: 0, bottom: 9, trailing: 0)
-
-    let groupSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(1), heightDimension: .estimated(455))
-    let group = NSCollectionLayoutGroup.vertical(layoutSize: groupSize, subitems: [item])
-
-    let section = NSCollectionLayoutSection(group: group)
-
-    return UICollectionViewCompositionalLayout(section: section)
-  }
-}
 
 // MARK: UICollectionViewDelegate
